Added a config option for auto-deletion. Also refined clean command.

+++ includes/common.js
@@ -2,25 +2,37 @@
 const sleep = require('sleep-promise');
 
 const logger = require('./logger.js');
+const app = require('../app.js');
 
 var msgs = exports.msgs = [];
 
 exports.sendMsg = function(msg,content,reply,timeout){
     let sent;
+    let push = true;
     if(reply){
-		msg.reply(content).then(message => cb(message));
+		msg.reply(content).then(message => cb(message));//.then(message => sent=message);//
     }else{
-		msg.channel.send(content).then(message => cb(message));
+		msg.channel.send(content).then(message => cb(message));//.then(message => sent=message);//
     }
-    if(!isNaN(timeout)){
+    if(!isNaN(timeout) && app.autoclean == 'yes'){
+    	push = false;
         sleep(timeout*1000).then(function(){
             sent.delete().catch(function(){logger.log('warn','Tried to delete a nonexistent message!')});
             msg.delete().catch(function(){logger.log('warn','Tried to delete a nonexistent message!')});
         })
     }
     function cb(message){
-    	sent = message;
-    	msgs.push(message);
-    	msgs.push(msg);
+    	if(push){
+	    	msgs.push(message);
+	    	msgs.push(msg);
+	    }
     }
+}
+
+exports.reset = function(){
+	msgs = [];
+}
+
+exports.getMsg = function(cb){
+	cb(msgs);
 }
